SkepTech Redesigned #7
Done
-Added Youtube Section
-made it responsive
-cleaned the old clutter form the 3d as well as the old projects

Next:
-Fix responsivensss
-Make a better copy
