+++ components/landingPage/contactus.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import GSAP from 'gsap'
 import { useEffect, useRef} from 'react'
-import Footer from './Footer'
+import Footer from '../Footer'
 
 const ContactUs = React.forwardRef<HTMLDivElement>((props, ref) => {
