+++ pages/index.js
@@ -3,17 +3,15 @@
 // import '../styles/App.scss'
 import Head from 'next/head'
 import Nav from '../components/nav'
-import Hero from '../components/hero'
-import Projects from '../components/projects'
-import About from '../components/about'
-import Services from '../components/services'
-import Contact from '../components/contactus'
+import Hero from '../components/landingPage/hero'
+import Projects from '../components/landingPage/projects'
+import About from '../components/landingPage/about'
+import Services from '../components/landingPage/services'
+import Contact from '../components/landingPage/contactus'
 import {useEffect, useRef, useState} from 'react'
-import ThreeCanvas from '../components/Canvas3d'
 import { useRouter } from 'next/router';
-import { useSpring } from 'react-spring';
 import { motion as m } from 'framer-motion'
-import Squares from '../components/squares'
+import Squares from '../components/landingPage/squares'
 
 // Custom hook to get the current page
 const useCurrentPage = () => {
@@ -29,9 +27,6 @@ const useCurrentPage = () => {
 };
 
 
-
-
-
 //Render Function
 
 const App = () => {
@@ -39,11 +34,6 @@ const App = () => {
   //Tab Categories props
   const [activeTab, setActiveTab] = useState(0);
   const [contentVisible, setContentVisible] = useState(true);
-  const [currentColor, setCurrentColor] = useState('#0ea7b5'); // Default color for Tab 1
-  const springProps = useSpring({
-    color: currentColor,
-    config: { duration: 500 }, // Adjust the duration as needed
-  });
 
   const handleTabClick = (index) => {
     setContentVisible(false); // Start the fade-out animation
@@ -51,21 +41,6 @@ const App = () => {
       setActiveTab(index);
       setContentVisible(true); // Start the fade-in animation
     }, 300); // Adjust the timeout based on your transition duration
-    
-    switch (index) {
-      case 0:
-        setCurrentColor('#0ea7b5');
-        break;
-      case 1:
-        setCurrentColor('#ff0000');
-        break;
-      case 2:
-        setCurrentColor('#00ff00');
-        break;
-      default:
-        setCurrentColor('#0ea7b5');
-        break;
-    }
   };
   
 //URL Updater
@@ -130,27 +105,6 @@ const App = () => {
   const containerRef = useRef(null)
 const currentPage = useCurrentPage();
 
- 
-  //Canvas Scroll Function
-  const [scrollTop, setScrollTop] = useState(0);
-  
-  const onScroll = () => {
-    const winScroll = document.documentElement.scrollTop;
-    const height =
-      document.documentElement.scrollHeight -
-      document.documentElement.clientHeight;
-  
-    const scrolled = (winScroll / height);
-  
-    setScrollTop(scrolled); // Update scrollTop first
-  };
-  
-  useEffect(() => {
-    window.addEventListener("scroll", onScroll);
-  
-    return () => window.removeEventListener("scroll", onScroll);
-  }, [scrollTop]);
-  
 
 
 
@@ -174,7 +128,6 @@ const currentPage = useCurrentPage();
         <link rel="preconnect" href="https://fonts.googleapis.com"></link>
         <link rel="preconnect" href="https://fonts.gstatic.com"></link>
         <link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Mono:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;1,100;1,200;1,300;1,400;1,500;1,600;1,700&display=swap" rel="stylesheet"></link>
-        {/* <ThreeCanvas scrollProgress={scrollTop} clr={springProps.color} /> */}
         <Squares 
           speed={0.3} 
           size={400} //pixels
@@ -209,9 +162,7 @@ const currentPage = useCurrentPage();
           <About ref={about} />
           <Services />
           <Contact ref={contact} />
-        </m.div>
-      {/* </LocomotiveScrollProvider> */}
-    
+        </m.div>    
     </m.div>
