+++ styles/App.scss
@@ -4,14 +4,6 @@
     color: $primary-color;
 }
 
-.canvas3d{
-    position: fixed;
-    left: 0;
-    top: 0;
-    height: 100vh;
-    width: 100vw;
-    z-index: -1;
-}
 .loader-wrapper{
     display: flex;
     align-items: center;
@@ -32,14 +24,4 @@ h1{
 .absoulte{
    
 }   
-
-.ThreeCanvas{
-    height: 100vh;
-    width: 100vw;
-    position: fixed;
-    top: 0 ;
-    left: 0;
-    z-index: -1;
-}
-
