+++ styles/index.scss
@@ -1,4 +1,4 @@
-@import "./variables.scss";
+@import "../styles/variables.scss";
 @import "./links.scss";
 @import "../styles/nav.scss";
 @import "../styles/hero.scss";
@@ -6,10 +6,10 @@
 @import "../styles/about.scss";
 @import "../styles/services.scss";
 @import "../styles/contactus.scss";
+@import "../styles/squares.scss";
 @import "../styles/projectPage.scss";
 @import "../styles/squares.scss";
 @import "../styles/Library.scss";
-
 @font-face {
     font-family: 'NR';
     src: url('../public/headingFont/Neue\ Regrade\ Variable.ttf') format('truetype');
@@ -17,6 +17,8 @@
 @import url('https://fonts.cdnfonts.com/css/helvetica-neue-9');
 @import "../node_modules/locomotive-scroll/src/locomotive-scroll.scss";
 @import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Mono:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;1,100;1,200;1,300;1,400;1,500;1,600;1,700&display=swap');
+
+
 html, body, div, span, applet, object, iframe,
 h1, h2, h3, h4, h5, h6, p, blockquote, pre,
 a, abbr, acronym, address, big, cite, code,
